Move vietneu_tts under community_models to match project contributions policy
