+++ include/engine/community_models/vietneu_tts/tokenizer_speech_encoder.h
@@ -3,8 +3,8 @@
 #include "engine/framework/core/execution_context.h"
 #include "engine/framework/assets/tensor_source.h"
 #include "engine/framework/runtime/session.h"
-#include "engine/models/vietneu_tts/assets.h"
-#include "engine/models/vietneu_tts/types.h"
+#include "engine/community_models/vietneu_tts/assets.h"
+#include "engine/community_models/vietneu_tts/types.h"
 
 #include <cstddef>
 #include <memory>
