+++ src/community_models/vietneu_tts/prompt_tts_custom_voice.cpp
@@ -1,4 +1,4 @@
-#include "engine/models/vietneu_tts/prompt_tts_custom_voice.h"
+#include "engine/community_models/vietneu_tts/prompt_tts_custom_voice.h"
 
 #include <stdexcept>
 #include <string>
