+++ src/community_models/vietneu_tts/tokenizer_speech_decoder.cpp
@@ -1,4 +1,4 @@
-#include "engine/models/vietneu_tts/tokenizer_speech_decoder.h"
+#include "engine/community_models/vietneu_tts/tokenizer_speech_decoder.h"
 
 #include "engine/framework/assets/tensor_source.h"
 #include "engine/framework/core/backend.h"
